Desarrollo

+++ controllers/EstudianteController.php
@@ -27,6 +27,8 @@ use app\models\Ciclo;
 use app\models\OpcionCurso;
 use app\models\ActaCalificacion;
 use app\models\User;
+use app\models\GrupoEstudianteSearch;
+use app\models\MateriaSearch;
 
 class EstudianteController extends Controller
 {
@@ -50,6 +52,36 @@ class EstudianteController extends Controller
                                 //Llamada al método que comprueba si es un administrador
                                 //return User::isUserAdministrador(Yii::$app->user->identity->idusuario);
                                 return User::isUserAutenticado(Yii::$app->user->identity->idusuario, 1);
+                            },  
+                        ],
+                        [
+                            //El administrador tiene permisos sobre las siguientes acciones
+                            'actions' => [''],//Especificar que acciones tiene permitidas este usuario
+                            //Esta propiedad establece que tiene permisos
+                            'allow' => true,
+                            //Usuarios autenticados, el signo ? es para invitados
+                            'roles' => ['@'],
+                            //Este método nos permite crear un filtro sobre la identidad del usuario
+                            //y así establecer si tiene permisos o no
+                            'matchCallback' => function ($rule, $action) {
+                                //Llamada al método que comprueba si es un administrador
+                                //return User::isUserAdministrador(Yii::$app->user->identity->idusuario);
+                                return User::isUserAutenticado(Yii::$app->user->identity->idusuario, 2);
+                            },  
+                        ],
+                        [
+                            //El administrador tiene permisos sobre las siguientes acciones
+                            'actions' => [''],//Especificar que acciones tiene permitidas este usuario
+                            //Esta propiedad establece que tiene permisos
+                            'allow' => true,
+                            //Usuarios autenticados, el signo ? es para invitados
+                            'roles' => ['@'],
+                            //Este método nos permite crear un filtro sobre la identidad del usuario
+                            //y así establecer si tiene permisos o no
+                            'matchCallback' => function ($rule, $action) {
+                                //Llamada al método que comprueba si es un administrador
+                                //return User::isUserAdministrador(Yii::$app->user->identity->idusuario);
+                                return User::isUserAutenticado(Yii::$app->user->identity->idusuario, 3);
                             },  
                         ],
                         [
@@ -436,23 +468,25 @@ class EstudianteController extends Controller
                          ->innerJoin(['cat_materias'], '`grupos`.`idmateria`=`cat_materias`.`idmateria`')
                          ->where(["grupos.idciclo" => $idciclo, "grupos_estudiantes.idestudiante" => $idestudiante])
                          ->all();
-        $creditos = $model1[0]['creditos'];
 
         $table2 = new \yii\db\Query();
         $model2 = $table2->from(['cat_carreras'])
-                         ->select(['cat_carreras.idcarrera','cat_carreras.desc_carrera'])
+                         ->select(['cat_carreras.idcarrera','cat_carreras.desc_carrera', 'estudiantes.nombre_estudiante'])
                          ->innerJoin(['estudiantes'], '`cat_carreras`.`idcarrera`=`estudiantes`.`idcarrera`')
                          ->where(["estudiantes.idestudiante" => $idestudiante])
                          ->all();
+
+        $creditos = $model1[0]['creditos'];
         $idcarrera = (count($model2) > 0) ? $model2[0]['idcarrera'] : "";
         $desc_carrera = (count($model2) > 0) ? $model2[0]['desc_carrera'] : "";
+        $estudiante = (count($model2) > 0) ? $model2[0]['nombre_estudiante'] : "";
 
         if(count($model) <= 0 && $status == 1)
         {
             $msg = "No se encontraron registros relacionados con el No. Control ". $idestudiante;
         }
 
-        return $this->render("horariomodificar", ["model" => $model, "form" => $form, "ciclos" => $ciclos, "idestudiante" => $idestudiante, "idciclo" => $idciclo, "idciclo_actual" => $idciclo_actual, "creditos" => $creditos, "idcarrera" => $idcarrera, "carrera" => $desc_carrera, "status" => $status, "msg" => $msg]);
+        return $this->render("horariomodificar", ["model" => $model, "form" => $form, "ciclos" => $ciclos, "idestudiante" => $idestudiante, "estudiante" => $estudiante, "idciclo" => $idciclo, "idciclo_actual" => $idciclo_actual, "creditos" => $creditos, "idcarrera" => $idcarrera, "carrera" => $desc_carrera, "status" => $status, "msg" => $msg]);
     }
 
     public function actionDeletehorarioestudiante()
@@ -462,18 +496,25 @@ class EstudianteController extends Controller
             $idestudiante = Html::encode($_GET["idestudiante"]);
             $idgrupo = Html::encode($_GET["idgrupo"]);
             $idciclo = Html::encode($_GET["idciclo"]);
-            /*
+
             $acta_calificacion = ActaCalificacion::find()
                                                  ->where(["idestudiante" => $idestudiante, "idgrupo" => $idgrupo])
                                                  ->count();
-            */
-            if(GrupoEstudiante::deleteAll(["idestudiante" => $idestudiante, "idgrupo" => $idgrupo]))
+
+            if($acta_calificacion == 0)
             {
-                $msg = "Registro eliminado";
+                if(GrupoEstudiante::deleteAll(["idestudiante" => $idestudiante, "idgrupo" => $idgrupo]))
+                {
+                    $msg = "Registro eliminado";
+                }
+                else
+                {
+                    $msg = "Error al eliminar el registro";
+                }
             }
             else
             {
-                $msg = "Error al eliminar el registro";
+                $msg = "El registro no puede ser eliminado, debido a que tiene información relacionada";
             }
 
             header("Location: ".Url::toRoute("/estudiante/horariomodificar?idestudiante=$idestudiante&idgrupo=$idgrupo&idciclo=$idciclo&msg=$msg&status=2"));
@@ -493,39 +534,34 @@ class EstudianteController extends Controller
 
             $opcion_curso = OpcionCurso::find()->orderBy(['desc_opcion_curso' => SORT_ASC])->all();
 
-            $query = "SELECT
-	                    grupos.idgrupo,
-	                    grupos.idmateria,
-	                    cat_materias.desc_materia,
-	                    cat_materias.creditos,
-	                    grupos.num_semestre,
-                        ciclo.desc_ciclo,
-                        grupos.desc_grupo
-                    FROM
-    	                grupos
-	                INNER JOIN cat_materias ON grupos.idmateria = cat_materias.idmateria
-                    INNER JOIN ciclo ON grupos.idciclo = ciclo.idciclo 
-                    WHERE
-    	                grupos.idcarrera = :idcarrera
-                    AND
-                        grupos.idciclo = :idciclo
-                    AND
-                        grupos.idgrupo NOT IN ((
-		                SELECT
-			                idgrupo
-		                FROM
-			                horario_estudiante_v 
-		                WHERE
-			                idestudiante = :idestudiante AND idciclo = :idciclo 
-		            ))
-                    ORDER BY
-	                    grupos.num_semestre DESC,
-	                    cat_materias.desc_materia ASC";
-            $materias = Yii::$app->db->createCommand($query)
-                                     ->bindValue(':idestudiante', $idestudiante)
-                                     ->bindValue(':idciclo', $idciclo)
-                                     ->bindValue(':idcarrera', $idcarrera)
-                                     ->queryAll();
+            $table = new \yii\db\Query();
+            $materias = $table->from(["grupos"])
+                              ->select([
+                                    "grupos.idgrupo",
+	                                "grupos.idmateria",
+        	                        "cat_materias.desc_materia",
+	                                "cat_materias.creditos",
+	                                "grupos.num_semestre",
+                                    "ciclo.desc_ciclo",
+                                    "grupos.desc_grupo"
+                                ])
+                                ->orderBy("cat_materias.desc_materia")
+                                ->innerJoin(["cat_materias"], "grupos.idmateria = cat_materias.idmateria")
+                                ->innerJoin(["ciclo"], "grupos.idciclo = ciclo.idciclo")
+                                ->where(["grupos.idcarrera" => $idcarrera])
+                                ->andFilterWhere(["grupos.idciclo" => $idciclo])
+                                ->andFilterWhere(["NOT IN", "grupos.idgrupo", (new \yii\db\Query())
+                                                                                    ->select('idgrupo')
+                                                                                    ->from('horario_estudiante_v')
+                                                                                    ->where(["idestudiante" => $idestudiante, "idciclo" => $idciclo])]);
+
+            if(Html::encode(isset($_GET["desc_materia"])))
+            {
+                $desc_materia = Html::encode($_GET["desc_materia"]);
+                $materias = $table->andFilterWhere(["LIKE", "cat_materias.desc_materia", $desc_materia]);
+            }
+
+            $materias = $table->all();
 
             return $this->render("horarioagregar", ["materias" => $materias, "opcion_curso" => $opcion_curso, "idestudiante" => $idestudiante, "idciclo" => $idciclo, "idcarrera" => $idcarrera]);
         }
@@ -563,5 +599,19 @@ class EstudianteController extends Controller
         }
 
         return $status;
+    }
+
+    public function actionGrid()
+    {
+        $this->layout = 'main2';
+
+        $searchModel = new GrupoEstudianteSearch();
+        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+
+        return $this->render("grid", [
+            "searchModel" => $searchModel,
+            "dataProvider" => $dataProvider
+        ]);
+        
     }
 }
